+++ examples/react/breadcrumbs/src/components/App.js
@@ -2,12 +2,12 @@ import React from "react";
 import { useResponse } from "@curi/react-dom";
 
 export default function App() {
-  const { response, router } = useResponse();
+  const { response } = useResponse();
 
   const { body: Body } = response;
   return (
     <div>
-      <Body response={response} router={router} />
+      <Body response={response} />
     </div>
   );
 }