+++ packages/react-universal/types/hooks/useResponse.d.ts
@@ -1,2 +1,2 @@
-import { Emitted } from "@curi/types";
-export default function useResponse(): Emitted;
+import { ResponseAndNav } from "@curi/types";
+export default function useResponse(): ResponseAndNav;