+++ website/src/pages/Tutorials/react-advanced.js
@@ -753,13 +753,15 @@ export default function Home({ response }) {
           array.
         </p>
 
-        <CodeBlock lang="jsx" data-line="7">
+        <CodeBlock lang="jsx" data-line="9">
           {`// src/components/Book.js
 import React from 'react';
+import { useRouter } from '@curi/react-dom';
 
 import cart from '../cart';
 
-export default function Book({ response, router }) {
+export default function Book({ response }) {
+  const router = useRouter();
   const { book } = response.data;
   if (!book) {
     return <div>The requested book could not be found</div>;